+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [1.4.10] - 2019-08-15
+### Changed
+- Fix a rename issue on windows
+
+## [1.4.9] - 2019-08-13
+### Changed
+- Update Built-By to be generic
+
+## [1.4.8] - 2019-08-13
+### Changed
+- Add option to force image ordering
+
+## [1.4.7] - 2019-08-13
+### Changed
+- Fixed a bug in image filename extensions 
+
 ## [1.4.6] - 2019-08-06
 ### Changed
 - Fixed a bug when creating image file names